Add several changes including status swipe delete
Used RecyclerView to implement the swipe functionality.  Statuses can be deleted by swiping from right to left on a list item.  This modifies the content displayed where the item was, changing the background color, showing a message to indicate the effect of the swipe, and displaying an undo button to revert the change.  If the undo button is not clicked within a given time then the item is removed and the deletion is made.

Added edit and delete functionality for contacts.  This includes code to prevent edits that do not provide the necessary information from being saved.  Currently this means that the edit will not be performed if the contact name is left empty.

A visible character limit is added to the the 'add status' dialog which updates as the status is modified.  In addition to this, a toast is displayed to the user to indicate an invalid status.  This is only called if the status is left empty at this point.

Implemented a style for dialog buttons

Deleted the StatusAdapter class as it wasn't being used.